Made yaml serializer standard
