+++ examples/debug.yaml
@@ -1,4 +1,4 @@
-# This is a super-small config just for debugging
+# Small config to help refactoring to make the model completely YAML based
 defaults:
   experiment:
     model_file: examples/output/<EXP>.mod
@@ -9,15 +9,34 @@ defaults:
     decode_every: 1
     eval_metrics: bleu,wer
   train:
-    train_src: examples/data/head.ja
-    train_trg: examples/data/head.en
-    dev_src: examples/data/head.ja
-    dev_trg: examples/data/head.en
     default_layer_dim: 64
     dropout: 0.5
-    encoder:
-      type: BiLSTM
-      dropout: 0.0
+    training_corpus: !BilingualTrainingCorpus
+      train_src: examples/data/head.ja
+      train_trg: examples/data/head.en
+      dev_src: examples/data/head.ja
+      dev_trg: examples/data/head.en
+    corpus_parser: !BilingualCorpusParser
+      src_reader: !PlainTextReader {}
+      trg_reader: !PlainTextReader {}
+    model: !DefaultTranslator
+      input_embedder: !SimpleWordEmbedder
+#        vocab_size: 100
+        emb_dim: 64
+      encoder: !BiLSTMEncoder
+        layers: 1
+        input_dim: 64
+      attender: !StandardAttender
+        state_dim: 64
+        hidden_dim: 64
+        input_dim: 64
+      output_embedder: !SimpleWordEmbedder
+#        vocab_size: 100
+        emb_dim: 64
+      decoder: !MlpSoftmaxDecoder
+        layers: 1
+#        vocab_size: 100
+        mlp_hidden_dim: 64
   decode:
     src_file: examples/data/head.ja
   evaluate:
@@ -29,9 +48,9 @@ debug-config-1layer:
 
 debug-config-2layers:
   train:
-    decoder_layers: 2
+    dropout: 0.2
 
 debug-config-2layers-finetune:
   train:
-    decoder_layers: 2
+    dropout: 0.2
     pretrained_model_file: examples/output/debug-config-2layers.mod
